code/requirements.txt
numpy
pandas
SQLAlchemy
#psycopg2 
psycopg2-binary #(Work for dlinux)